Add experiment models
Single table inheritance was chosen since `type` was specified as a
required attribute on the `experiments` table. This makes it easy to
define behavior on typed models later, for example adding special class
or instance methods to FieldExperiment or LabExperiment, while still
having all experiments live in a single table. We get the benefit of not
having to define multiple validations on each type, such as enforcing
unique names across all experiments, regardless of type.