Add roles through STI on users
- Create STI relationship for user roles
- Generate predicates for checking user role

Note: By using STI we have to keep in mind two things that Rails will no
longer take care of. First, we have to manually update the type attribute
on objects when the type changes from User. Second, we need to update
the object in-memory by using User.becomes(Experimenter), for example.
If we assign types upon user creation, this won't be necessary. For
example, using Experimenter.new(...) instead of User.new(...).